Update and enable the remaining mail opcodes

+++ src/server/game/Entities/Player/Player.cpp
@@ -3961,7 +3961,7 @@ void Player::SendNewMail()
 {
     // deliver undelivered mail
     WorldPacket data(SMSG_RECEIVED_MAIL, 4);
-    data << (uint32) 0;
+    data << float(0.0);
     GetSession()->SendPacket(&data);
 }
